portals: get holiday date importer back
Refactored ImportFileController
