+++ library/IvozProvider/Klear/Importer/ImporterControllerTrait.php
@@ -1,0 +1,430 @@
+<?php
+namespace IvozProvider\Klear\Importer;
+
+trait ImporterControllerTrait
+{
+    // Traits cannot have constants
+    protected $MYSQL_DUPLICATE_ENTRY_ERROR_CODE = 1062;
+
+    protected $_mainRouter;
+    protected $_item;
+    protected $_pk;
+    protected $_parentPk;
+    protected $_ids;
+    protected $_mapperName;
+    protected $_mapper;
+    protected $_modelName;
+    protected $_model;
+    protected $_availableFields;
+    protected $_multilangFields;
+    protected $_availableLangs;
+    protected $_config;
+    protected $_freeUploadCommand;
+    protected $_forcedValues;
+
+    protected $_csvParser;
+    protected $_genericFormBuilder;
+    protected $_customFormBuilder;
+
+    public function init()
+    {
+        /**
+         * Initialize action controller here
+         */
+        if ((!$this->_mainRouter = $this->getRequest()->getUserParam("mainRouter")) ||
+            (!is_object($this->_mainRouter)) ) {
+            throw new \Zend_Exception("", Zend_Controller_Plugin_ErrorHandler::EXCEPTION_NO_ACTION);
+        }
+
+        $this->_item    = $this->_mainRouter->getCurrentItem();
+        $this->_config = $this->_item->getConfig();
+        $this->_buildCsvParser();
+
+        $this->_mapperName = $this->_config->getProperty("mapper");
+        $this->_freeUploadCommand = $this->_config->getProperty("freeUploadCommand");
+
+        $this->_mapper = new $this->_mapperName;
+
+        $this->_pk = $this->getRequest()->getParam("pk");
+        $this->_parentPk = $this->getRequest()->getParam("parentId");
+
+        $this->_model = $this->_item->getModelSpec()->getInstance();
+        $this->_modelName = $this->_item->getModelSpec()->getClassName();
+        $availableFields = $this->_item->getModelSpec()->getFields();
+
+        $this->_multilangFields = $this->_model->getMultiLangColumnsList();
+        $this->_availableLangs = $this->_model->getAvailableLangs();
+        asort($this->_availableLangs);
+
+        $this->_forcedValues = $this->_config->getProperty("forcedValues")->toArray();
+
+        $this->_availableFields = array();
+
+        foreach ($availableFields as $key => $field) {
+            $this->_availableFields[$key] = $field;
+            if (isset($this->_multilangFields[$key])) {
+                $title = \Klear_Model_Gettext::gettextCheck($field->get("title"));
+                foreach ($this->_availableLangs as $lang) {
+                    $newField = $field->toArray();
+                    $newField["title"] = $title." ".ucfirst($lang);
+                    $newFieldConfig = new \Zend_Config($newField);
+                    $this->_availableFields[$key."_".$lang] = $newFieldConfig;
+                }
+            }
+        }
+
+        $selectableFields = array_diff_key($this->_availableFields, $this->_forcedValues);
+        $this->_genericFormBuilder = new GenericForm(
+            $selectableFields,
+            $this->_getFilteredPostValues(),
+            $this->_helper->translate
+        );
+
+        $this->_helper->layout->disableLayout();
+    }
+
+    protected function _buildCsvParser()
+    {
+        $delimiter = $this->_config->getProperty("delimiter");
+        $enclosure = $this->_config->getProperty("enclosure");
+        $scape = $this->_config->getProperty("scape");
+
+        $this->_csvParser = new CsvParser($delimiter, $enclosure, $scape);
+    }
+
+    protected function _dispatch(array $data)
+    {
+        $jsonResponse = new \Klear_Model_DispatchResponse();
+        $jsonResponse->setModule('klearMatrix');
+        $jsonResponse->setPlugin('klearMatrixGenericDialog');
+        $jsonResponse->addJsFile("/js/plugins/qq-fileuploader.js");
+        $jsonResponse->addJsFile('/js/plugins/jquery.klearmatrix.genericdialog.js');
+        $jsonResponse->setData($data);
+        $jsonResponse->attachView($this->view);
+    }
+
+    protected function _loadFile($message = null)
+    {
+
+        if (!$message) {
+            $message .= "<p>".$this->_helper->translate("Please, choose a file to import")."</p>";
+        }
+        $message .= "<div style='text-align:center;'><input type='file' name='importFile' data-upload-command='" . $this->_freeUploadCommand . "'></div>";
+
+        $data = array(
+            'title' => $this->_helper->translate("Parse file and import"),
+            'message' => $message,
+            'buttons' => array(
+                $this->_helper->translate('Close') => array(
+                    'recall' => false,
+                    'reloadParent' => false
+                ),
+                $this->_helper->translate('Continue') => array(
+                    'recall' => true,
+                    'params' => array(
+                        'step' => "load_file"
+                    )
+                ),
+            )
+        );
+        $this->_dispatch($data);
+    }
+
+    protected function _getFilePath()
+    {
+        $tempFSystemNS = new \Zend_Session_Namespace('File_Controller');
+
+        $fileHash = $this->getRequest()->getParam('importFile');
+        if (!empty($fileHash) && isset($tempFSystemNS->{$fileHash})) {
+            $tempFile = $tempFSystemNS->{$fileHash};
+            return $tempFile["path"];
+        }
+        return null;
+    }
+
+    protected function _parseFile($filePath, $limit = null)
+    {
+        $lines = $this->_csvParser->parseFile($filePath);
+
+        if (count($lines)<=0) {
+            return false;
+        }
+
+        $counter = 0;
+        $linesArray = array();
+        foreach ($lines as $lineFields) {
+
+            $linesArray[] = $lineFields;
+            $counter++;
+            if (!is_null($limit) && $counter==$limit) {
+                break;
+            }
+        }
+
+        return $linesArray;
+    }
+
+    protected function _buildForm($lines)
+    {
+        if ($this->_customFormBuilder) {
+            return $this->_customFormBuilder->getForm($lines);
+        }
+
+        return $this->_genericFormBuilder->getForm($lines);
+    }
+
+    protected function _getFilteredPostValues()
+    {
+        $values = [];
+
+        $params = $this->getRequest()->getParams();
+        foreach ($params as $key => $value) {
+
+            if (false === strpos($key, 'field_')) {
+                continue;
+            }
+            $values[$key] = $value;
+        }
+
+        $values['ingoreFirst'] = $this->getRequest()->getParam('ingoreFirst', false);
+        $values['update'] = $this->getRequest()->getParam("update", false);
+
+        return $values;
+    }
+
+    protected function _confirmImport($message = "", $filePath = null)
+    {
+        if (is_null($filePath)) {
+            $filePath = $this->_getFilePath();
+        }
+        $lines = $this->_parseFile($filePath, 3);
+
+        if ($lines===false) {
+            $this->_generalError();
+            return;
+        }
+
+        $form = $this->_buildForm($lines);
+        $data = array(
+            'title' => $this->_helper->translate("Parse file and import"),
+            'message' => $message."<br>".$form,
+            'buttons' => array(
+                $this->_helper->translate('Close') => array(
+                    'recall' => false,
+                    'reloadParent' => false
+                ),
+                $this->_helper->translate('Yes, continue!') => array(
+                    'recall' => true,
+                    'params' => array(
+                        "step" => 'process_confirm',
+                        "filePath" => $filePath
+                    )
+                ),
+            )
+        );
+        $this->_dispatch($data);
+    }
+
+    protected function _import()
+    {
+        $this->_log("Importing file");
+        $error = array();
+        $success = array();
+        $filePath = $this->getRequest()->getParam('filePath');
+        $ignoreFirst = $this->getRequest()->getParam('ingoreFirst');
+        $conf = array();
+
+        $lines = $this->_parseFile($filePath, 1);
+        $fileFields = count($lines[0]);
+        $fieldsPossitions = array();
+        for($i = 0; $i < $fileFields; $i++) {
+            $fieldName = $this->getRequest()->getParam('field_'.$i);
+            $fieldsPossitions[$fieldName] = $i;
+        }
+
+        $result = $this->_save($lines, $fieldsPossitions, $filePath);
+        $message = $result["message"];
+        $this->_log($message);
+
+        if ($result["error"]) {
+            $this->_confirmImport($message, $filePath);
+            return;
+        }
+
+        $data = array (
+            'title' => $this->_helper->translate("Parse file and import"),
+            'message' => $message,
+            'buttons' => array(
+                $this->_helper->translate('Close') => array(
+                    'recall' => false,
+                    'reloadParent' => true
+                )
+            )
+        );
+        $this->_dispatch($data);
+    }
+
+    protected function _save($lines, $fieldsPossitions, $filePath = "")
+    {
+        //comprobar campos requeridos
+        $missing = false;
+        $message = '<p><b style="color: red;">ERROR</b></p>';
+        foreach ($this->_availableFields as $key => $field) {
+            if (isset($this->_forcedValues[$key])) {
+                continue;
+            }
+            if ($field->get("required")) {
+                if (!isset($fieldsPossitions[$key])) {
+                    $missing = true;
+                    $message .= "<p>'".$key."' ".$this->_helper->translate("is a required field").".</p>";
+                }
+            }
+        }
+
+        if ($missing) {
+            $this->_log($message);
+            return array("error" => true, "message" => $message);
+        }
+
+        $calculateHighestPk = false;
+        $message = $this->_helper->translate("File imported");
+        $this->_log("Opening file..");
+        $firstLine = true;
+        $handle = fopen($filePath, "r");
+        if ($handle) {
+            $errors = array();
+            while (!feof($handle)) {
+                $buffer = fgets($handle, 4096);
+                if (!$buffer) {
+                    continue;
+                }
+                $buffer = trim($buffer);
+                if (empty($buffer)) {
+                    continue;
+                }
+
+                $lineFields = str_getcsv($buffer,";");
+                if ($this->getRequest()->getParam("ingoreFirst") && $firstLine){
+                    $firstLine = false;
+                    $this->_log("Ignoring first line");
+                    continue;
+                }
+
+                $data = array();
+                foreach ($fieldsPossitions as $fieldName => $fieldPosition) {
+                    $fieldPosition = intval($fieldPosition);
+                    if ($fieldName == "ignore") {
+                        continue;
+                    }
+                    $value = $lineFields[$fieldPosition];
+
+                    if(empty($value)) {
+                        $value = "";
+                    }
+
+                    if (strtolower($value) == "null") {
+                        $value = null;
+                    }
+
+                    $data[$fieldName] = $value;
+                }
+
+                $conditions = array();
+                if (isset($data["id"])) {
+                    $conditions["id"] = $data["id"];
+                }
+                foreach ($this->_forcedValues as $field => $value) {
+                    if ($value == "%parent%") {
+                        $value = $this->_parentPk;
+                    }
+                    $data[$field] = $value;
+                    $conditions[$field] = $value;
+                }
+
+                try {
+                    if ($this->getRequest()->getParam("update", false) && isset($data["id"])) {
+                        $model = $this->_mapper->findOneByField(array_keys($conditions), array_values($conditions));
+                        if (!is_null($model)) {
+                            $model->populateFromArray($data);
+                            $model->save();
+                        } else {
+                            $wheres = array();
+                            foreach ($conditions as $key => $value) {
+                                $wheres[] = $key." = '".$value."'";
+                            }
+                            $error = "No model found. where ".implode(" AND ", $wheres);
+                            $this->_log($error);
+                            $errors[] = $error;
+                        }
+                    } else {
+                        $model = new $this->_modelName();
+                        $model->populateFromArray($data);
+                        $model->save(true);
+                    }
+                } catch (\Exception $e) {
+                    if ($e->getCode() == $this->MYSQL_DUPLICATE_ENTRY_ERROR_CODE) {
+                        $calculateHighestPk = true;
+                    }
+                    $this->_log($e->getMessage());
+                    $errors[] = $e->getMessage();
+                }
+            }
+        }
+        fclose($handle);
+
+        if ($calculateHighestPk) {
+            $maxIdModel = $this->_mapper->fetchOne(null, "id desc");
+            $errors[] = "Highest id is '".$maxIdModel->getPrimaryKey()."'";
+        }
+        if (!empty($errors)) {
+            $message .= " " . $this->_helper->translate("but following errors ocurred"). ":<ul>";
+
+            for ($i = 0; $i < count($errors); $i++) {
+                if ($i > 3) {
+                    $message .= "<li>... </li>";
+                    break;
+                }
+
+                $error = $errors[$i];
+                $message .= "<li>".$this->_helper->translate($error)."</li>";
+            }
+
+            $message .= "</ul>";
+        }
+        return array ("error" => false, "message" => $message);
+    }
+
+    protected function _generalError()
+    {
+        $this->_dispatch(
+            array(
+                'title' => $this->_helper->translate("An error occurred"),
+                'message' => $this->_helper->translate("An error occurred: check files format and content."),
+                'buttons' => array(
+                    $this->_helper->translate('Close') => array(
+                        'recall' => false,
+                        'reloadParent' => false
+                    )
+                )
+            )
+        );
+    }
+
+    protected function _log($message, $priorityText = null)
+    {
+        switch ($priorityText) {
+            case "error":
+                $priority = \Zend_Log::ERR;
+                break;
+            case "info":
+                $priority = \Zend_Log::INFO;
+                break;
+            default:
+                $priority = \Zend_Log::DEBUG;
+                break;
+        }
+
+        $message = '[File Importer] ' . $message;
+        $this->_helper->log($message, $priority);
+    }
+}
